Struct bitm::ArrayWithRankSelect101111

pub struct ArrayWithRankSelect101111<Select = BinaryRankSearch, Select0 = BinaryRankSearch> {
    pub content: Box<[u64]>,
    pub l1ranks: Box<[usize]>,
    pub l2ranks: Box<[u64]>,
    /* private fields */
}

The structure that holds array of bits content and ranks structure that takes no more than 3.125% extra space. It can return the number of ones (or zeros) in first index bits of the content (see rank and rank0 method) in O(1) time. In addition, it supports select queries utilizing binary search over ranks (see BinaryRankSearch) or (optionally, at the cost of about 0.39% extra space overhead) combined sampling (which is usually faster; see CombinedSampling).

It uses modified version of the structure described in the paper:

Zhou D., Andersen D.G., Kaminsky M. (2013) “Space-Efficient, High-Performance Rank and Select Structures on Uncompressed Bit Sequences”. In: Bonifaci V., Demetrescu C., Marchetti-Spaccamela A. (eds) Experimental Algorithms. SEA 2013. Lecture Notes in Computer Science, vol 7933. Springer, Berlin, Heidelberg. https://doi.org/10.1007/978-3-642-38527-8_15

The modification consists of different level 2 entries that hold 4 rank values (r0 <= r1 <= r2 <= r3) relative to level 1 entry. The content of level 2 entry, listing from the least significant bits, is:

original: r0 stored on 32 bits, r1-r0 on 10 bits, r2-r1 on 10 bits, r3-r2 on 10 bits;

our: r0 stored on 32 bits, r3-r0 on 11 bits, r2-r0 on 11 bits, r1-r0 on 10 bits (and unused fields in the last entries, for out-of-bound content bits, are filled with bit ones). With this layout, we can read the corresponding value in the rank query without branching and avoid some bound checks in the select query.

For in-word selection, it uses the select64 function.

Fields
content: Box<[u64]>l1ranks: Box<[usize]>l2ranks: Box<[u64]>
Implementations

impl<S: SelectForRank101111, S0: Select0ForRank101111> ArrayWithRankSelect101111<S, S0>

pub fn build(content: Box<[u64]>) -> (Self, usize)