Struct bitm::ArrayWithRankSimple

pub struct ArrayWithRankSimple {
    pub content: Box<[u64]>,
    pub ranks: Box<[u32]>,
}

The structure that holds array of bits content and ranks structure that takes no more than 6.25% extra space. It can returns the number of ones in first index bits of the content (see rank method) in O(1) time. Only content with less than 2^32 bit ones is supported.

Usually ArrayWithRankSelect101111 should be preferred to ArrayWithRankSimple.

Fields
content: Box<[u64]>ranks: Box<[u32]>
Implementations

impl ArrayWithRankSimple

pub fn build(content: Box<[u64]>) -> (Self, u32)

Constructs ArrayWithRankSimple and count number of bits set in content. Returns both.

pub fn try_rank(&self, index: usize) -> Option<u32>

pub fn rank(&self, index: usize) -> u32