Trait bitm::Select0

pub trait Select0 {
    // Required method
    fn try_select0(&self, rank: usize) -> Option<usize>;

    // Provided methods
    fn select0(&self, rank: usize) -> usize { ... }
    unsafe fn select0_unchecked(&self, rank: usize) -> usize { ... }
}

Trait implemented by the types that support select zero queries, i.e. can (quickly) find the position of the n-th zero in the bitmap.

Required Methods

fn try_select0(&self, rank: usize) -> Option<usize>

Returns the position of the rank-th zero (counting from 0) in self or None if there are no such many zeros in self.

Provided Methods

fn select0(&self, rank: usize) -> usize

Returns the position of the rank-th zero (counting from 0) in self or panics if there are no such many zeros in self.

unsafe fn select0_unchecked(&self, rank: usize) -> usize

Returns the position of the rank-th zero (counting from 0) in self. The result is undefined if there are no such many zeros in self.